BJS2-15628
- Вынесена логика по проверки прав в отдельный сервис валидатор
- Правки по замечаниям
- Поправлены тесты и дописан тест на класс валидатор

+++ src/main/java/faang/school/projectservice/dto/campaign/CampaignDto.java
@@ -3,8 +3,6 @@ package faang.school.projectservice.dto.campaign;
 import faang.school.projectservice.dto.client.Currency;
 import faang.school.projectservice.model.CampaignStatus;
 
-import java.math.BigDecimal;
-
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
@@ -28,10 +26,10 @@ public class CampaignDto {
 
     @Min(1)
     @NotNull
-    private BigDecimal goal;
+    private Double goal;
 
     @Min(1)
-    private BigDecimal amountRaised;
+    private Double amountRaised;
 
     @NotNull
     private CampaignStatus status;
